funcionando toda la seguridad

+++ src/main/java/com/meadowspace/meadowSpaceProject/services/img/UploadFilesService.java
@@ -64,9 +64,10 @@ public class UploadFilesService implements IUploadFilesService {
 	public String deleteFile(String nameFile) throws IOException {
 		Path path = Paths.get(FOLDER, nameFile);
 
-//		if (!Files.exists(path)) {
+		if (!Files.exists(path)) {
 //			throw new FileNotFoundException("El archivo " + nameFile + " no existe en la carpeta especificada");
-//		}
+			return null;
+		}
 
 		Files.delete(path);
